# openapi.yml
openapi: 3.0.2
info:
  title: Rich Internet Applications
  version: '1.0'
...
paths:
  /api/users/verify-email:
    post:
      tags:
        - user
      summary: Send reset password email
      operationId: sendResetPasswordEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmail"
      responses:
        '200':
          description: A verification email has been sent
        '400':
          description: Failed to read body
  /api/users/reset-password:
    post:
      tags:
        - user
      summary: Reset password
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPassword"
      responses:
        '200':
          description: Successfully reset password
        '400':
          description: Failed to read body
...
  /api/users/verificationMail:
    post:
      tags:
        - user
      summary: Send verification email for registration
      operationId: sendVerificationMail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        '200':
          description: Check mail
        '400':
          description: Failed to read body or Account with that email already exists!
...
  /api/users/reset-superadmin-password:
    post:
      tags:
        - user
      summary: Reset superadmin password
      operationId: resetSuperAdminPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordInput"
      responses:
        '200':
          description: Password changed successfully
        '400':
          description: Failed to read body or Admin not found or Error from database!
        '403':
          description: User is not authorized to perform this action
  /api/users/add-admin:
    post:
      tags:
        - user
      summary: Add admin
      operationId: addAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Admin"
      responses:
        '200':
          description: Successfully added admin
        '400':
          description: Failed to read body or There is already an account with the entered email
        '403':
          description: User is not authorized to perform this action
  /api/users/edit-admin:
    post:
      tags:
        - user
      summary: Edit superadmin profile
      operationId: editSuperAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Admin"
      responses:
        '200':
          description: Successfully edit profile
        '400':
          description: Failed to read body or Database error
        '403':
          description: User is not authorized to perform this action
...
components:
  schemas:
    VerifyEmail:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPassword:
      type: object
      properties:
        password:
          type: string
        token:
          type: string
      required:
        - password
        - token
...
    RegisterInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        surname:
          type: string
        role:
          type: integer
      required:
        - email
        - password
        - name
        - surname
        - role
...
    PasswordInput:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    Admin:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
      required:
        - name
        - surname
        - email
...